Handle more case for first order dangling objects

+++ octopoes/octopoes/core/service.py
@@ -178,8 +178,14 @@ class OctopoesService:
                 logger.debug("Affirmation source %s already deleted", origin.source)
                 return
 
-        if origin.origin_type == OriginType.AFFIRMATION and not self.origin_repository.list_origins(
-            valid_time=valid_time, result=origin.source
+        if origin.origin_type == OriginType.AFFIRMATION and not any(
+            origin
+            for origin in self.origin_repository.list_origins(
+                origin_type={OriginType.DECLARATION, OriginType.OBSERVATION, OriginType.INFERENCE},
+                valid_time=valid_time,
+                result=origin.source,
+            )
+            if not (origin.origin_type == OriginType.INFERENCE and [origin.source] == origin.result)
         ):
             logger.debug("Affirmation source %s seems dangling, deleting", origin.source)
             self.ooi_repository.delete(origin.source, valid_time)
